change input parameter from int to float.

+++ Synth/src/streams/SynthesizerStream.h
@@ -11,9 +11,9 @@ namespace Synth {
 	public:
 		SynthesizerStream();
 		~SynthesizerStream();
-		void addSine(int f);
-		void addSqrt(int f);
-		void addTrig(int f);
+		void addSine(float f);
+		void addSqrt(float f);
+		void addTrig(float f);
 		virtual float tick();
 
 	private:
